+++ app/lib/widget/user_dialog/user_details.dart
@@ -74,7 +74,7 @@ class UserDetailsGroup extends StatelessWidget {
               notifier.userHeight = newHeight;
             }
           },
-          onEditingComplete: () {
+          onSubmitted: () {
             onRefresh();
           },
         ),