+++ CHANGELOG.md
@@ -17,6 +17,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Bugfix 🐛:
 - Fixed the height field in the personalization settings losing focus after
   every digit, which made it impossible to enter a height in one go
+- Fixed the done key of the keyboard doing nothing in the height field
 
 
 ## [1.3.0] - 2026-08-26